requirements.txt
python-telegram-bot==21.6
websockets==12.0
